+++ tests/Catalog/SearchProductAttributeTest.php
@@ -18,23 +18,24 @@ final class SearchProductAttributeTest extends TestCase
     {
         $jsonData = <<<JSON
 {
-	"attribute": {
-		"id": 52,
-		"name": "Couleur test",
-		"type": "M",
-		"position": 10,
-		"parentId": null,
-		"code": ""
-	},
-	"values": [{
-		"id": 64,
-		"attributeId": 52,
-		"name": "blue",
-		"slug": "blue-color",
-		"image": {
-			"id": 708
-		}
-	}]
+    "attribute": {
+        "id": 52,
+        "name": "Couleur test",
+        "type": "M",
+        "position": 10,
+        "parentId": null,
+        "code": ""
+    },
+    "values": [{
+        "id": 64,
+        "attributeId": 52,
+        "name": "blue",
+        "slug": "blue-color",
+        "image": {
+            "id": 708,
+            "url": {}
+        }
+    }]
 }
 JSON;
 
@@ -52,6 +53,7 @@ JSON;
                 'slug' => 'blue-color',
                 'image' => [
                     'id' => 708,
+                    'url' => [],
                 ],
             ]),
         ], $attribute->getValues());